requirements.txt
torch>=1.11.0
torchvision>=0.12.0
gradio>=4.44.0
numpy
pillow
matplotlib
seaborn
scikit-learn